requirements.txt
google-cloud-storage
google-cloud-batch
python-dotenv